djangoapp/requirements.txt
Django>=5.1,<5.2
djangorestframework>=3.14.0
django_cors_headers>=4.4.0
Pillow>=9.0.0
psycopg2-binary>=2.9.9,<2.10
pytest-django>=4.8.0,<5.0
pytest>=8.3.2,<8.4